streamlit_app/requirements.txt
streamlit
google-generativeai
pandas
plotly
numpy
